Implement multithreading by creating multiple instances of Gramadon reader
if one reader stops, all readers are stopped and the program terminates

Details:
  - garmadon reader creation has been moved to a dedicated factory class
  - adapt GarmadonReader to support many instances
  - adapt PrometheusMetrics to support multiple KafkaConsumers
