+++ readers/hdfs/src/test/java/com/criteo/hadoop/garmadon/hdfs/offset/HeartbeatConsumerTest.java
@@ -25,7 +25,7 @@ public class HeartbeatConsumerTest {
         final HeartbeatConsumer hb = new HeartbeatConsumer<>(writers, Duration.ofMillis(10));
 
         hb.handle(buildGarmadonMessage(new TopicPartitionOffset(TOPIC, BASE_PARTITION, BASE_OFFSET)));
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
         Thread.sleep(500);
 
         hb.stop().join();
@@ -51,7 +51,7 @@ public class HeartbeatConsumerTest {
         hb.handle(buildGarmadonMessage(secondPartitionSecondOffset));
         hb.handle(buildGarmadonMessage(secondPartitionFirstOffset));
 
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
         Thread.sleep(1000);
 
         for (PartitionedWriter<String> writer: writers) {
@@ -75,7 +75,7 @@ public class HeartbeatConsumerTest {
         final Offset secondOffset = new TopicPartitionOffset(TOPIC, BASE_PARTITION + 1, BASE_OFFSET + 11);
 
         hb.handle(buildGarmadonMessage(firstOffset));
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
         Thread.sleep(1000);
         verify(writer, times(1)).heartbeat(eq(BASE_PARTITION),
                 argThat(new OffsetArgumentMatcher(firstOffset)));
@@ -99,7 +99,7 @@ public class HeartbeatConsumerTest {
         hb.handle(buildGarmadonMessage(secondPartitionOffset));
         hb.dropPartition(BASE_PARTITION);
 
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
 
         Thread.sleep(1000);
 
@@ -121,7 +121,7 @@ public class HeartbeatConsumerTest {
         final HeartbeatConsumer hb = new HeartbeatConsumer<>(writers, Duration.ofMillis(10));
         hb.handle(buildGarmadonMessage(offset));
 
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
 
         verify(writers.get(0), after(100).atLeast(1)).heartbeat(eq(PARTITION),
                 argThat(new OffsetArgumentMatcher(offset)));
@@ -134,7 +134,7 @@ public class HeartbeatConsumerTest {
         final HeartbeatConsumer hb = new HeartbeatConsumer<String>(Collections.singleton(mock(PartitionedWriter.class)),
                 Duration.ofMillis(10));
 
-        hb.start(mock(Thread.UncaughtExceptionHandler.class));
+        hb.start(mock(Thread.UncaughtExceptionHandler.class), "heartbeat");
         Thread.sleep(1000);
         hb.stop().join();
     }
