+++ readers/hdfs/src/test/java/com/criteo/hadoop/garmadon/hdfs/writer/PartitionedWriterTest.java
@@ -362,7 +362,7 @@ public class PartitionedWriterTest {
         final PartitionedWriter.Expirer<String> expirer = new PartitionedWriter.Expirer<>(
                 Arrays.asList(firstConsumer, secondConsumer), Duration.ofMillis(1));
 
-        expirer.start(mock(Thread.UncaughtExceptionHandler.class));
+        expirer.start(mock(Thread.UncaughtExceptionHandler.class), "expirer");
 
         Thread.sleep(500);
 
@@ -379,7 +379,7 @@ public class PartitionedWriterTest {
         final PartitionedWriter.Expirer<String> expirer = new PartitionedWriter.Expirer<>(Collections.emptyList(),
                 Duration.ofMillis(10));
 
-        expirer.start(mock(Thread.UncaughtExceptionHandler.class));
+        expirer.start(mock(Thread.UncaughtExceptionHandler.class), "expirer");
         Thread.sleep(500);
         expirer.stop().join();
     }
@@ -389,7 +389,7 @@ public class PartitionedWriterTest {
         final PartitionedWriter.Expirer<String> expirer = new PartitionedWriter.Expirer<>(
                 Collections.singleton(mock(PartitionedWriter.class)), Duration.ofHours(42));
 
-        expirer.start(mock(Thread.UncaughtExceptionHandler.class));
+        expirer.start(mock(Thread.UncaughtExceptionHandler.class), "expirer");
         Thread.sleep(1000);
         expirer.stop().join();
     }
